+++ tests/geometryRTheta/advection_field_rtheta/advection_field_gtest.cpp
@@ -28,12 +28,12 @@
 #include "geometry_r_theta.hpp"
 #include "l_norm_tools.hpp"
 #include "mesh_builder.hpp"
+#include "polar_foot_finder.hpp"
 #include "quadrature.hpp"
 #include "rk3.hpp"
 #include "rk4.hpp"
 #include "simulation_utils_tools.hpp"
 #include "spline_definitions_r_theta.hpp"
-#include "spline_polar_foot_finder.hpp"
 #include "spline_quadrature.hpp"
 #include "test_cases_adv_field.hpp"
 #include "trapezoid_quadrature.hpp"
@@ -147,11 +147,12 @@ TEST(AdvectionFieldRThetaComputation, TestAdvectionFieldFinder)
             theta_extrapolation_rule);
 
     RK3Builder const time_stepper;
-    SplinePolarFootFinder find_feet(
-            grid,
+    PolarFootFinder find_feet = make_polar_foot_finder<
+            FootFindingSpace::PSEUDO_PHYSICAL,
+            AdvectionFieldSpace::PHYSICAL>(
             time_stepper,
             to_physical_mapping,
-            to_physical_mapping,
+            grid,
             builder,
             spline_evaluator_extrapol);