+++ tests/geometryRTheta/advection_rtheta/advection_selected_test.cpp
@@ -29,35 +29,16 @@
 #include "mesh_builder.hpp"
 #include "paraconfpp.hpp"
 #include "params.yaml.hpp"
+#include "polar_foot_finder.hpp"
 #include "polar_spline_evaluator.hpp"
 #include "rk2.hpp"
 #include "rk3.hpp"
 #include "rk4.hpp"
 #include "spline_definitions_r_theta.hpp"
-#include "spline_polar_foot_finder.hpp"
 
 
 
 namespace fs = std::filesystem;
-
-namespace {
-#if defined(CIRCULAR_MAPPING_PHYSICAL)
-using X_adv = X;
-using Y_adv = Y;
-#elif defined(CZARNY_MAPPING_PHYSICAL)
-using X_adv = X;
-using Y_adv = Y;
-
-#elif defined(CZARNY_MAPPING_PSEUDO_CARTESIAN)
-using X_adv = X_pC;
-using Y_adv = Y_pC;
-
-#elif defined(DISCRETE_MAPPING_PSEUDO_CARTESIAN)
-using X_adv = X_pC;
-using Y_adv = Y_pC;
-#endif
-
-} //end namespace
 
 int main(int argc, char** argv)
 {
@@ -172,45 +153,25 @@ int main(int argc, char** argv)
     CircularToCartesian<R, Theta, X, Y> to_physical_mapping;
     CircularToCartesian<R, Theta, X, Y> to_physical_mapping_host;
     CartesianToCircular<X, Y, R, Theta> to_logical_analytical_mapping;
-    CircularToCartesian<R, Theta, X, Y> const& logical_to_pseudo_cart_mapping(
-            to_physical_analytical_mapping);
     std::string const mapping_name = "CIRCULAR";
-    std::string const adv_domain_name = "PHYSICAL";
-    key += "circular_physical";
+    key += "circular";
 #else
 
     double const czarny_e = 0.3;
     double const czarny_epsilon = 1.4;
 
-#if defined(CZARNY_MAPPING_PHYSICAL)
-    using LogicalToPseudoPhysicalMapping = CzarnyToCartesian<R, Theta, X, Y>;
     CzarnyToCartesian<R, Theta, X, Y> to_physical_analytical_mapping(czarny_e, czarny_epsilon);
+    CartesianToCzarny<X, Y, R, Theta> to_logical_analytical_mapping(czarny_e, czarny_epsilon);
+
+#if defined(CZARNY_MAPPING_PHYSICAL) || defined(CZARNY_MAPPING_PSEUDO_CARTESIAN)
     CzarnyToCartesian<R, Theta, X, Y> to_physical_mapping(czarny_e, czarny_epsilon);
     CzarnyToCartesian<R, Theta, X, Y> to_physical_mapping_host(czarny_e, czarny_epsilon);
-    CartesianToCzarny<X, Y, R, Theta> to_logical_analytical_mapping(czarny_e, czarny_epsilon);
-    CzarnyToCartesian<R, Theta, X, Y> const& logical_to_pseudo_cart_mapping(
-            to_physical_analytical_mapping);
     std::string const mapping_name = "CZARNY";
-    std::string const adv_domain_name = "PHYSICAL";
-    key += "czarny_physical";
+    key += "czarny";
 
-#elif defined(CZARNY_MAPPING_PSEUDO_CARTESIAN)
-    using LogicalToPseudoPhysicalMapping = CircularToCartesian<R, Theta, X_pC, Y_pC>;
-    CzarnyToCartesian<R, Theta, X, Y> to_physical_analytical_mapping(czarny_e, czarny_epsilon);
-    CzarnyToCartesian<R, Theta, X, Y> to_physical_mapping(czarny_e, czarny_epsilon);
-    CzarnyToCartesian<R, Theta, X, Y> to_physical_mapping_host(czarny_e, czarny_epsilon);
-    CartesianToCzarny<X, Y, R, Theta> to_logical_analytical_mapping(czarny_e, czarny_epsilon);
-    CircularToCartesian<R, Theta, X_pC, Y_pC> logical_to_pseudo_cart_mapping;
-    std::string const mapping_name = "CZARNY";
-    std::string const adv_domain_name = "PSEUDO CARTESIAN";
-    key += "czarny_pseudo_cartesian";
-
-#else
-#if not defined(DISCRETE_MAPPING_PSEUDO_CARTESIAN)
+#elif not defined(DISCRETE_MAPPING_PSEUDO_CARTESIAN)
     static_assert(false, "No mapping macro defined");
-#endif
-    CzarnyToCartesian<R, Theta, X, Y> to_physical_analytical_mapping(czarny_e, czarny_epsilon);
-    CartesianToCzarny<X, Y, R, Theta> to_logical_analytical_mapping(czarny_e, czarny_epsilon);
+#else
     DiscretePoloidalCSSplineMappingBuilder<
             X,
             Y,
@@ -234,11 +195,20 @@ int main(int argc, char** argv)
                     spline_evaluator_extrapol);
     using LogicalToPseudoPhysicalMapping = CircularToCartesian<R, Theta, X_pC, Y_pC>;
     DiscretePoloidalCSSplineMapping to_physical_mapping = mapping_builder();
-    CircularToCartesian<R, Theta, X_pC, Y_pC> logical_to_pseudo_cart_mapping;
     std::string const mapping_name = "DISCRETE";
-    std::string const adv_domain_name = "PSEUDO CARTESIAN";
-    key += "discrete_pseudo_cartesian";
+    key += "discrete";
 #endif
+#endif
+
+#if defined(CIRCULAR_MAPPING_PHYSICAL) || defined(CZARNY_MAPPING_PHYSICAL)
+    constexpr FootFindingSpace FFSpace = FootFindingSpace::PHYSICAL;
+    std::string const adv_domain_name = "PHYSICAL";
+    key += "_physical";
+
+#elif defined(CZARNY_MAPPING_PSEUDO_CARTESIAN) || defined(DISCRETE_MAPPING_PSEUDO_CARTESIAN)
+    constexpr FootFindingSpace FFSpace = FootFindingSpace::PSEUDO_PHYSICAL;
+    std::string const adv_domain_name = "PSEUDO CARTESIAN";
+    key += "_pseudo_cartesian";
 #endif
 
     key += "-";
@@ -291,21 +261,13 @@ int main(int argc, char** argv)
         fs::create_directory(output_folder);
     }
 
-    using X_adv = typename LogicalToPseudoPhysicalMapping::cartesian_tag_x;
-    using Y_adv = typename LogicalToPseudoPhysicalMapping::cartesian_tag_y;
-    using PseudoCartesianToCircular = CartesianToCircular<X_adv, Y_adv, R, Theta>;
-    using PseudoPhysicalToPhysicalMapping
-            = CombinedMapping<decltype(to_physical_mapping), PseudoCartesianToCircular>;
-    PseudoPhysicalToPhysicalMapping
-            pseudo_physical_to_physical(to_physical_mapping, PseudoCartesianToCircular(), 1e-12);
-
-    SplinePolarFootFinder const foot_finder(
-            grid,
-            time_stepper,
-            pseudo_physical_to_physical,
-            logical_to_pseudo_cart_mapping,
-            builder,
-            spline_evaluator_extrapol);
+    PolarFootFinder const foot_finder
+            = make_polar_foot_finder<FFSpace, AdvectionFieldSpace::PHYSICAL>(
+                    time_stepper,
+                    to_physical_mapping,
+                    grid,
+                    builder,
+                    spline_evaluator_extrapol);
 
     BslAdvectionPolar
             advection_operator(builder, spline_evaluator, foot_finder, to_physical_mapping);
@@ -314,9 +276,7 @@ int main(int argc, char** argv)
               << " - " << simu_type << " : " << std::endl;
     simulate(
             to_physical_mapping_host,
-            to_physical_mapping,
             to_logical_analytical_mapping,
-            logical_to_pseudo_cart_mapping,
             to_physical_analytical_mapping,
             grid,
             foot_finder,