+++ CHANGELOG.md
@@ -31,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add a MPI scope guard.
 - Add MI250X toolchain for LUMI.
 - Add a CMake option to control the discovery timeout of Gyselalib++ tests.
+- Added `PolarFootFinder` to replace `SplinePolarFootFinder` for clarity of Advection and Foot finding spaces.
 
 ### Fixed
 
@@ -72,9 +73,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Use Simpson quadrature for XYVxVy Landau damping.
 - Update py-dask in the CPU Spack toolchain.
 - Disable python for PDI and pdiplugin-pycall on macOS.
-- In `SplinePolarFootFinder`, interpolate advection field on the provided coordinate system.
+- Enforce version of Kokkos Tools in all toolchains for reproducibility.
+- Remove default O1 optimisation flag.
 
 ### Deprecated
+
+- Deprecated `SplinePolarFootFinder`.
 
 ### Removed